# src/main/resources/distance-openapi.yml
openapi: 3.0.3
info:
  title: Distance
  description: |-
    This api updates a given postcode's coordinates. It also calculate the distance between 2 postcodes using Haversine's formula.
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080/v1
tags:
  - name: postcode
    description: Description about postcode
paths:
  /postcode/{postcode}:
    put:
      tags:
        - postcode
      description: Update an existing postcode
      operationId: updatePostcode
      parameters:
        - name: postcode
          in: path
          required: true
          schema:
            type: string
            example: AB10 1XG
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostcode'
      responses:
        '200':
          description: Success
        '400':
          description: Invalid Postcode
    get:
      tags:
        - postcode
      description: Gets an existing postcode
      parameters:
        - name: postcode
          in: path
          required: true
          schema:
            type: string
            example: AB10 1XG
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Postcode'
        '400':
          description: Invalid Postcode
  /postcode/distance:
    post:
      tags:
        - postcode
      description: Calculates the distance between 2 postcodes. Uses Haversine formula.
      operationId: calculateDistance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateDistance'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Distance'
        '400':
          description: Invalid Postcode
components:
  schemas:
    UpdatePostcode:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          example: 57.1441560
        longitude:
          type: number
          format: double
          example: -2.1148640
    CalculateDistance:
      type: object
      required:
        - postcodeA
        - postcodeB
      properties:
        postcodeA:
          type: string
          example: AB10 1XG
        postcodeB:
          type: string
          example: AB10 1XG
    Distance:
      type: object
      properties:
        distance:
          type: number
          format: double
          example: km
        unit:
          type: string
        postcodeA:
          type: object
          $ref: '#/components/schemas/Postcode'
        postcodeB:
          type: object
          $ref: '#/components/schemas/Postcode'
    Postcode:
      type: object
      properties:
        postcode:
          type: string
          example: AB10 1XG
        latitude:
          type: number
          format: double
          example: 57.1441560
        longitude:
          type: number
          format: double
          example: -2.1148640